Rename the Bean Validation extension to Hibernate Validator
See #591.
