+++ bom/pom.xml
@@ -34,7 +34,7 @@
 
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-bean-validation-deployment</artifactId>
+            <artifactId>shamrock-hibernate-validator-deployment</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
          </dependency>
