+++ core/deployment/src/main/java/org/jboss/shamrock/deployment/builditem/FeatureBuildItem.java
@@ -24,7 +24,7 @@ public final class FeatureBuildItem extends MultiBuildItem {
     public static final String MP_REACTIVE_OPERATORS = "mp-reactive-streams-operators";
     public static final String MP_REACTIVE_MESSAGING = "mp-reactive-streams-messaging";
     public static final String REACTIVE_CONVERTERS = "reactive-type-converters";
-    public static final String BEAN_VALIDATION = "bean-validation";
+    public static final String HIBERNATE_VALIDATOR = "hibernate-validator";
     public static final String JPA = "jpa";
     public static final String SCHEDULER = "scheduler";
     public static final String SECURITY = "security";
