--- extensions/bean-validation/runtime/src/main/java/org/jboss/shamrock/beanvalidation/runtime/ArcProxyBeanMetaDataClassNormalizer.java
@@ -1,20 +1,0 @@
-package org.jboss.shamrock.beanvalidation.runtime;
-
-import org.hibernate.validator.metadata.BeanMetaDataClassNormalizer;
-import org.jboss.protean.arc.Subclass;
-
-/**
- * In the case of a proxy generated by Arc, return the parent class.
- */
-public class ArcProxyBeanMetaDataClassNormalizer implements BeanMetaDataClassNormalizer {
-
-    @Override
-    public Class<?> normalize(Class<?> clazz) {
-        if (Subclass.class.isAssignableFrom(clazz)) {
-            return clazz.getSuperclass();
-        }
-
-        return clazz;
-    }
-
-}
