--- extensions/bean-validation/runtime/src/main/java/org/jboss/shamrock/beanvalidation/runtime/ValidatorHolder.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-package org.jboss.shamrock.beanvalidation.runtime;
-
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-
-public class ValidatorHolder {
-
-    private static ValidatorFactory validatorFactory;
-
-    private static Validator validator;
-
-    static void initialize(ValidatorFactory validatorFactory) {
-        ValidatorHolder.validatorFactory = validatorFactory;
-        ValidatorHolder.validator = validatorFactory.getValidator();
-    }
-
-    static ValidatorFactory getValidatorFactory() {
-        return validatorFactory;
-    }
-
-    static Validator getValidator() {
-        return validator;
-    }
-}
