--- extensions/bean-validation/runtime/src/main/java/org/jboss/shamrock/beanvalidation/runtime/ValidatorProvider.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.shamrock.beanvalidation.runtime;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-
-@ApplicationScoped
-public class ValidatorProvider {
-
-    @Produces
-    public ValidatorFactory factory() {
-        return ValidatorHolder.getValidatorFactory();
-    }
-
-    @Produces
-    public Validator validator() {
-        return ValidatorHolder.getValidator();
-    }
-}
