--- extensions/bean-validation/runtime/src/main/java/org/jboss/shamrock/beanvalidation/runtime/interceptor/MethodValidated.java
@@ -1,20 +1,0 @@
-package org.jboss.shamrock.beanvalidation.runtime.interceptor;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.interceptor.InterceptorBinding;
-
-/**
- * Marker class to indicate a method should be validated.
- */
-@Inherited
-@InterceptorBinding
-@Retention(value = RetentionPolicy.RUNTIME)
-@Target({ ElementType.TYPE, ElementType.METHOD })
-public @interface MethodValidated {
-
-}
