--- extensions/bean-validation/runtime/src/main/java/org/jboss/shamrock/beanvalidation/runtime/jaxrs/JaxrsEndPointValidated.java
@@ -1,20 +1,0 @@
-package org.jboss.shamrock.beanvalidation.runtime.jaxrs;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.interceptor.InterceptorBinding;
-
-/**
- * Marker class to indicate a JAX-RS end point should be validated.
- */
-@Inherited
-@InterceptorBinding
-@Retention(value = RetentionPolicy.RUNTIME)
-@Target({ ElementType.TYPE, ElementType.METHOD })
-public @interface JaxrsEndPointValidated {
-
-}
