+++ extensions/hibernate-validator/runtime/src/main/java/org/jboss/shamrock/hibernate/validator/runtime/ArcProxyBeanMetaDataClassNormalizer.java
@@ -1,0 +1,20 @@
+package org.jboss.shamrock.hibernate.validator.runtime;
+
+import org.hibernate.validator.metadata.BeanMetaDataClassNormalizer;
+import org.jboss.protean.arc.Subclass;
+
+/**
+ * In the case of a proxy generated by Arc, return the parent class.
+ */
+public class ArcProxyBeanMetaDataClassNormalizer implements BeanMetaDataClassNormalizer {
+
+    @Override
+    public Class<?> normalize(Class<?> clazz) {
+        if (Subclass.class.isAssignableFrom(clazz)) {
+            return clazz.getSuperclass();
+        }
+
+        return clazz;
+    }
+
+}
