+++ extensions/hibernate-validator/runtime/src/main/resources/META-INF/services/javax.ws.rs.ext.Providers
