--- integration-tests/bean-validation/src/main/java/org/jboss/shamrock/example/beanvalidation/BeanValidationApplication.java
@@ -1,9 +1,0 @@
-package org.jboss.shamrock.example.beanvalidation;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/bean-validation")
-public class BeanValidationApplication extends Application {
-
-}
