--- integration-tests/bean-validation/src/main/java/org/jboss/shamrock/example/beanvalidation/GreetingService.java
@@ -1,13 +1,0 @@
-package org.jboss.shamrock.example.beanvalidation;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.validation.constraints.NotNull;
-
-@ApplicationScoped
-public class GreetingService {
-
-    public String greeting(@NotNull String name) {
-        return "hello " + name;
-    }
-
-}
