--- integration-tests/bean-validation/src/main/java/org/jboss/shamrock/example/beanvalidation/custom/MyOtherBean.java
@@ -1,19 +1,0 @@
-package org.jboss.shamrock.example.beanvalidation.custom;
-
-@MyCustomConstraint
-public class MyOtherBean {
-
-    private String name;
-
-    public MyOtherBean(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
