--- integration-tests/bean-validation/src/test/java/org/jboss/shamrock/example/test/BeanValidationFunctionalityInGraalITCase.java
@@ -1,11 +1,0 @@
-package org.jboss.shamrock.example.test;
-
-import org.jboss.shamrock.test.junit.SubstrateTest;
-
-/**
- * Test various Bean Validation operations running in SubstrateVM
- */
-@SubstrateTest
-public class BeanValidationFunctionalityInGraalITCase extends BeanValidationFunctionalityTest {
-
-}
