+++ integration-tests/hibernate-validator/src/main/java/org/jboss/shamrock/example/MainMethod.java
