+++ integration-tests/hibernate-validator/src/main/java/org/jboss/shamrock/example/hibernate/validator/GreetingService.java
@@ -1,0 +1,13 @@
+package org.jboss.shamrock.example.hibernate.validator;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.validation.constraints.NotNull;
+
+@ApplicationScoped
+public class GreetingService {
+
+    public String greeting(@NotNull String name) {
+        return "hello " + name;
+    }
+
+}
