+++ integration-tests/hibernate-validator/src/main/java/org/jboss/shamrock/example/hibernate/validator/custom/MyOtherBean.java
@@ -1,0 +1,19 @@
+package org.jboss.shamrock.example.hibernate.validator.custom;
+
+@MyCustomConstraint
+public class MyOtherBean {
+
+    private String name;
+
+    public MyOtherBean(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
