+++ integration-tests/hibernate-validator/src/main/resources/ValidationMessages.properties
