+++ integration-tests/hibernate-validator/src/test/java/org/jboss/shamrock/example/test/HibernateValidatorFunctionalityInGraalITCase.java
@@ -1,0 +1,11 @@
+package org.jboss.shamrock.example.test;
+
+import org.jboss.shamrock.test.junit.SubstrateTest;
+
+/**
+ * Test various Bean Validation operations running in SubstrateVM
+ */
+@SubstrateTest
+public class HibernateValidatorFunctionalityInGraalITCase extends HibernateValidatorFunctionalityTest {
+
+}
