+++ integration-tests/main/pom.xml
@@ -79,7 +79,7 @@
         </dependency>
         <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-bean-validation-deployment</artifactId>
+            <artifactId>shamrock-hibernate-validator-deployment</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
